Add initial implementation.
This plugin is intended to provide charting functionality for Hudson CI with a set of advanced
features which will be described below. Actually this plugin is a fork of hudson-plot-plugin
originally written by somebody from Yahoo. At this moment following changes are made on original
codebase:
 * Refactoring: most of code has been removed and simplified.
 * Unit tests for series stuff added.
 * Big classes rename occured *Plot -> *Graph. It done in order to match new plugin name.
Also I would like to describe planned feature set:
 * No server side chart rendering anymore. It should be done at client side.
 * Zooming and panning support.
 * Auto resize support when browser window is resized.
 * Area, bar, line and variouse combinations of this types support at the same chart.
 * Logscale support.
But following feature set of original plugin is going to be removed (but this list is a subject
of change, I mean listed features can be implemented on demand):
 * 3d charting support.
 * floating labels at series point with custom urls.
Unfortunately client chart generation is not implemented at this moment because research about
appropriate javascript charting library is still in progress.